Reorganize constants to seperate folder
