+++ src/constants/errorCorrection.ts
