+++ src/constants/galois.ts
