+++ test/index.test.ts
@@ -1,3 +1,5 @@
+import QRious from "../src/QRious"
+
 test("woo", () => {
   expect(1 + 2).toBe(3)
 })
